# shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/members": {
      "post": {
        "operationId": "MembersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: members"
        ]
      },
      "get": {
        "operationId": "MembersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: members"
        ]
      }
    },
    "/admin/members/{id}": {
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: members"
        ]
      }
    },
    "/admin/members/{id}/skills": {
      "post": {
        "operationId": "MembersController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSkillDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: members"
        ]
      }
    },
    "/admin/skills": {
      "post": {
        "operationId": "SkillsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSkillDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: skills"
        ]
      },
      "get": {
        "operationId": "SkillsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SkillResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: skills"
        ]
      }
    },
    "/admin/skills/{id}": {
      "delete": {
        "operationId": "SkillsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin: skills"
        ]
      }
    },
...
  },
  "info": {
    "title": "SprintWell API",
    "description": "Sprint planning backend — admin, member and public endpoints.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice@example.com"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Alice"
          },
          "role": {
            "enum": [
              "MEMBER",
              "ADMIN"
            ],
            "type": "string",
            "example": "MEMBER"
          },
          "initialPassword": {
            "type": "string",
            "minLength": 8,
            "example": "changeme123"
          }
        },
        "required": [
          "email",
          "name",
          "role",
          "initialPassword"
        ]
      },
      "MemberResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role"
        ]
      },
      "AssignSkillDto": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "minLength": 1,
            "example": "skill-uuid"
          },
          "level": {
            "type": "number",
            "minimum": 1,
            "maximum": 5,
            "example": 3
          }
        },
        "required": [
          "skillId",
          "level"
        ]
      },
      "CreateSkillDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "example": "Python"
          }
        },
        "required": [
          "name"
        ]
      },
      "SkillResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
...
    }
  }
}